LearnerParam {ParamHelpers}	R Documentation

Create a description object for a parameter of a machine learning algorithm.

Description

This specializes Param by adding a few more attributes, like a default value, whether it refers to a training or a predict function, etc.

Usage

  makeNumericLearnerParam(id, lower = -Inf, upper = Inf,
    default, when = "train", requires = NULL)

  makeNumericVectorLearnerParam(id, len = as.integer(NA),
    lower = -Inf, upper = Inf, default, when = "train",
    requires = NULL)

  makeIntegerLearnerParam(id, lower = -Inf, upper = Inf,
    default, when = "train", requires = NULL)

  makeIntegerVectorLearnerParam(id, len = as.integer(NA),
    lower = -Inf, upper = Inf, default, when = "train",
    requires = NULL)

  makeDiscreteLearnerParam(id, values, default,
    when = "train", requires = NULL)

  makeDiscreteVectorLearnerParam(id, len = as.integer(NA),
    values, default, when = "train", requires = NULL)

  makeLogicalLearnerParam(id, default, when = "train",
    requires = NULL)

  makeLogicalVectorLearnerParam(id, len = as.integer(NA),
    default, when = "train", requires = NULL)

  makeUntypedLearnerParam(id, default, when = "train",
    requires = NULL)

  makeFunctionLearnerParam(id, default, when = "train",
    requires = NULL)

Arguments

id
[character(1)]
See Param.

len
[integer(1)]
See Param.

lower
[numeric]
See Param.

upper
[numeric]
See Param.

values
[vector | list]
See Param.

requires
[NULL | R expression]
See Param.

default
[any]
Default value used in learner. If this argument is missing, it means no default value is available.

when
[character(1)]
Specifies when parameter is used in the learner: “train”, “predict” or “both”. Default is “train”.

Details

The S3 class is a Param which additionally stores these elements:

default [any]

See argument of same name.

has.default [logical(1)]

Was a default value provided?

when [character(1)]

See argument of same name.

Value

[LearnerParam].

[Package ParamHelpers version 1.2 Index]