Param {ParamHelpers}	R Documentation

Create a description object for a parameter.

Description

For each parameter type a special constructor function is available, see below.

Usage

  makeNumericParam(id, lower = -Inf, upper = Inf,
    trafo = NULL, requires = NULL)

  makeNumericVectorParam(id, len, lower = -Inf,
    upper = Inf, trafo = NULL, requires = NULL)

  makeIntegerParam(id, lower = -Inf, upper = Inf,
    trafo = NULL, requires = NULL)

  makeIntegerVectorParam(id, len, lower = -Inf,
    upper = Inf, trafo = NULL, requires = NULL)

  makeLogicalParam(id, requires = NULL)

  makeLogicalVectorParam(id, len, requires = NULL)

  makeDiscreteParam(id, values, requires = NULL)

  makeDiscreteVectorParam(id, len, values, requires = NULL)

  makeFunctionParam(id, requires = NULL)

  makeUntypedParam(id, requires = NULL)

Arguments

id
[character(1)]
Name of parameter.

len
[integer(1)]
Length of vector parameter.

lower
[numeric]
Lower bounds. A singe value of length 1 is automatically replicated to len for vector parameters. Default is -Inf.

upper
[numeric]
Upper bounds. A singe value of length 1 is automatically replicated to len for vector parameters. Default is Inf.

values
[vector | list]
Possible discrete values. Instead of using a vector of atomic values, you are also allowed to pass a list of quite “complex” R objects, which are used as discrete choices. If you do the latter, the elements must be uniquely named, so that the names can be used as internal represenatations for the choice.

trafo
[NULL | function(x)]
Function to transform parameter. It should be applied to the parameter value before it is, e.g., passed to a corresponding objective function. Function must accept a parameter value as the first argument and return a transformed one. Default is NULL which means no transformation.

requires
[NULL | R expression]
States requirements on other paramaters' values, so that setting this parameter only makes sense if its requirements are satisfied (dependent parameter). Only really useful if the parameter is included in a ParamSet. Default is NULL which means no requirements.

Details

The S3 class is a list which stores these elements:

id [character(1)]

See argument of same name.

type [character(1)]

Data type of parameter. Possible types are “numeric”, “numericvector”, “integer”, “integervector”, “logical”, “logicalvector”, “discrete”, “discretevector”, “function”, “untyped”.

len [integer(1)]

See argument of same name.

lower [numeric]

See argument of same name. Length of this vector is len.

upper [numeric]

See argument of same name. Length of this vector is len.

values [list]

Discrete values, always stored as a named list.

trafo [NULL | function(x)]

See argument of same name.

requires [NULL | expression]

See argument of same name.

Value

[Param].

Examples

makeNumericParam("x",lower=-1, upper=1)

## Num param 'x'. Constr: -1 to 1. Trafo: FALSE. Requires: FALSE

makeNumericVectorParam("x", len=2)

## Num vec param 'x'. Len: 2. Constr: -Inf,-Inf to Inf,Inf. Trafo: FALSE. Requires: FALSE

makeDiscreteParam("y", values=c("a","b"))

## Disc param 'y'. Vals: a,b. Trafo: FALSE. Requires: FALSE

[Package ParamHelpers version 1.2 Index]