Convert optimization path to data.frame.

The following types of columns are created:

x-numeric(vector)	numeric

x-integer(vector)	integer

x-discrete(vector)	factor (names of values = levels)

x-logical(vector)	logical

y-columns	numeric

dob	integer

eol	integer

error.message	character

exec.time	numeric

extra-columns	any

If you want to convert these, look at BBmisc::convertDataFrameCols(). Dependent parameters whose constraints are unsatisfied generate NA entries in their respective columns. Factor columns of discrete parameters always have their complete level set from the param.set.

# S3 method for OptPathDF
as.data.frame(
  x,
  row.names = NULL,
  optional = FALSE,
  include.x = TRUE,
  include.y = TRUE,
  include.rest = TRUE,
  dob = x$env$dob,
  eol = x$env$eol,
  ...
)

Arguments

x

(OptPath())
Optimization path.

row.names

character
Row names for result. Default is none.

optional

(any)
Currently ignored.

include.x

(logical(1))
Include all input params? Default is TRUE.

include.y

(logical(1))
Include all y-columns? Default is TRUE.

include.rest

(logical(1))
Include all other columns? Default is TRUE.

dob

integer
Vector of date-of-birth values to further subset the result. Only elements with a date-of-birth included in dob are selected. Default is all.

eol

integer
Vector of end-of-life values to further subset the result. Only elements with an end-of-life included in eol are selected. Default is all.

...

(any)
Currently ignored.

Value

data.frame.