convertParamSetToIrace {ParamHelpers}	R Documentation

Converts a ParamSet object to a parameter object of the irace package.

Description

Converts to a textual description used in irace and then potentially calls readParameters.

Usage

convertParamSetToIrace(par.set, digits = 4, as.chars = FALSE)

Arguments

par.set
[ParamSet]
Parameter set.

digits
[integer(1)]
Number of digits used in transformation to irace textual description, codeformatC is used here internally.

as.chars
[logical(1)]
Return results as character vector of lines FALSE or call readParameters on it (TRUE). Default is FALSE.

Value

[list].

[Package ParamHelpers version 1.2 Index]