dfRowsToList {ParamHelpers}	R Documentation

Convert a data.frame row to list of parameter-value-lists.

Description

Please note that (naturally) the columns of df have to be of the correct type type w.r.t. the corresponding parameter. The only exception are integer parameters where the corresponding columns in df are allowed to be numerics. Dependent parameters whose requirements are not satisfied are represented by a scalar NA in the output.

Usage

  dfRowsToList(df, par.set)

  dfRowToList(df, par.set, i)

Arguments

df
[data.frame]
Data.frame, potentially from OptPathDF. Columns are assumed to be in the same order as par.set.

par.set
[ParamSet]
Parameter set.

i
[integer(1)]
Row index.

Value

[list]. Named by parameter ids.

[Package ParamHelpers version 1.2 Index]