Generates a design with the defaults of a parameter set.

The following types of columns are created:

numeric(vector)	numeric

integer(vector)	integer

discrete(vector)	factor (names of values = levels)

logical(vector)	logical

This will create a design containing only one point at the default values of the supplied param set. In most cases you will combine the resulting data.frame with a different generation function e.g. generateDesign(), generateRandomDesign() or generateGridDesign(). This is useful to force an evaluation at the default location of the parameters while still generating a design. Parameters default values, whose conditions (requires) are not fulfilled will be set to NA in the result.

generateDesignOfDefaults(par.set, trafo = FALSE)

Arguments

par.set
ParamSet
Parameter set.

trafo
(logical(1))
Transform all parameters by using theirs respective transformation functions. Default is FALSE.

Value

data.frame. Columns are named by the ids of the parameters. If the par.set argument contains a vector parameter, its corresponding column names in the design are the parameter id concatenated with 1 to dimension of the vector. The result will have an logical(1) attribute “trafo”, which is set to the value of argument trafo.