getLower {ParamHelpers}	R Documentation

Get lower / upper bounds and allowed discrete values for parameters.

Description

getLower and getUpper return a numerical vector of lower and upper bounds, getValues returns a list of possible value sets for discrete parameters.

Usage

  getLower(par.set, with.nr = FALSE)

  getUpper(par.set, with.nr = FALSE)

  getValues(par.set)

Arguments

par.set
[ParamSet]
Parameter set.

with.nr
[logical(1)]
Should number from 1 to length be appended to names of vector params? Default is FALSE.

Details

Parameters for which such bound make no sense - due to their type - are not present in the result.

Value

[vector | list]. Named by parameter ids.

Examples

ps <- makeParamSet(
  makeNumericParam("u"),
  makeIntegerParam("v", lower=1, upper=2),
  makeDiscreteParam("w", values=1:2),
  makeNumericVectorParam("x", len=2, lower=c(0, 10), upper=c(1, 11))
)
getLower(ps)

##    u    v    x    x 
## -Inf    1    0   10

getUpper(ps)

##   u   v   x   x 
## Inf   2   1  11

ps <- makeParamSet(
  makeNumericParam("u"),
  makeDiscreteParam("v", values=c("a", "b")),
  makeDiscreteParam("w", values=list(a=list(), b=NULL))
)
getValues(ps)

## $v
## $v$a
## [1] "a"
## 
## $v$b
## [1] "b"
## 
## 
## $w
## $w$a
## list()
## 
## $w$b
## NULL

[Package ParamHelpers version 1.2 Index]