getOptPathEl {ParamHelpers}	R Documentation

Get an element from the optimization path.

Description

Dependent parameters whose requirements are not satisfied are represented by a scalar NA in the elements of x of the return value.

Usage

  getOptPathEl(op, index)

Arguments

op
[OptPath]
Optimization path.

index
[integer(1)]
Index of element.

Value

List with elements x [named list], y [named numeric], dob [integer(1)] and eol [integer(1)].

[Package ParamHelpers version 1.2 Index]