getOptPathParetoFront {ParamHelpers}	R Documentation

Get indices of pareto front of optimization path.

Description

Get indices of pareto front of optimization path.

Usage

getOptPathParetoFront(op, y.names = op$y.names, dob = op$env$dob,
  eol = op$env$eol, index = FALSE)

Arguments

op
[OptPath]
Optimization path.

y.names
[character]
Names of performance measures to construct pareto front for. Default is all performance measures.

dob
[integer]
Possible dates of birth to select elements from. Defaults to all.

eol
[integer]
Possible end of life to select elements from. Defaults to all.

index
[logical(1)]
Return indices into path of front or y-matrix of nondominated points? Default is FALSE.

Value

[matrix | integer]. Either matrix (with named columns) of points of front in objective space or indices into path for front.

See Also

Other optpath: OptPath, OptPathDF, makeOptPathDF; addOptPathEl; getOptPathBestIndex; getOptPathCol; getOptPathDOB; getOptPathEOL; getOptPathEl; getOptPathErrorMessages; getOptPathExecTimes; getOptPathLength; getOptPathY; setOptPathElDOB; setOptPathElEOL

Examples

ps = makeParamSet(makeNumericParam("x"))
op = makeOptPathDF(par.set = ps, y.names = c("y1", "y2"), minimize = c(TRUE, TRUE))
addOptPathEl(op, x = list(x = 1), y = c(5, 3))
addOptPathEl(op, x = list(x = 2), y = c(2, 4))
addOptPathEl(op, x = list(x = 3), y = c(9, 4))
addOptPathEl(op, x = list(x = 4), y = c(4, 9))
as.data.frame(op)

##   x y1 y2 dob eol
## 1 1  5  3   1  NA
## 2 2  2  4   2  NA
## 3 3  9  4   3  NA
## 4 4  4  9   4  NA

getOptPathParetoFront(op)

## Loading required package: emoa

##   y1 y2
## 1  5  3
## 2  2  4

getOptPathParetoFront(op, index = TRUE)

## [1] 1 2

[Package ParamHelpers version 1.2 Index]