Get data.frame of input points (X-space) referring to the param set from the optimization path.

Get data.frame of input points (X-space) referring to the param set from the optimization path.

getOptPathX(op, dob, eol)

Arguments

op

OptPath
Optimization path.

dob

integer
Vector of date-of-birth values to further subset the result. Only elements with a date-of-birth included in dob are selected. Default is all.

eol

integer
Vector of end-of-life values to further subset the result. Only elements with an end-of-life included in eol are selected. Default is all.

Value

data.frame.

See also