getParamTypeCounts {ParamHelpers}	R Documentation

Returns information on the number of parameters of a each type.

Description

Returns information on the number of parameters of a each type.

Usage

getParamTypeCounts(par.set)

Arguments

par.set
[ParamSet]
Parameter set.

Value

[list] Named list which contains for each supported parameter type the number of parameters of this type in the given ParamSet.

[Package ParamHelpers version 1.2 Index]