Returns type information for a parameter set.

Returns type information for a parameter set.

getParamTypes(
  par.set,
  df.cols = FALSE,
  df.discretes.as.factor = TRUE,
  use.names = FALSE,
  with.nr = TRUE
)

Arguments

par.set
ParamSet
Parameter set.

df.cols
(logical(1))
If FALSE simply return the parameter types in the set, i.e., par$type. If TRUE, convert types so they correspond to R types of a data.frame where values of this set might be stored. This also results in replication of output types for vector parameters. Default is FALSE.

df.discretes.as.factor
(logical(1))
If df.cols is TRUE: Should type for discrete params be factor or character? Default is TRUE.

use.names
(logical(1))
Name the result vector? Default is FALSE.

with.nr
(logical(1))
Should number from 1 to length be appended to name? Only used if use.names and df.cols are TRUE. Default is TRUE.

Value

character.