getTypes {ParamHelpers}	R Documentation

Returns type information for a parameter set.

Description

Returns type information for a parameter set.

Usage

  getTypes(par.set, df.cols = FALSE, use.names = FALSE,
    with.nr = TRUE)

Arguments

par.set
[ParamSet]
Parameter set.

df.cols
[logical(1)]
If FALSE simply return the parameter types in the set, i.e., par$type. If TRUE, convert types so they correspond to R types of a data.frame where values of this set might be stored. This also results in replication of output types for vector parameters. Default is FALSE.

use.names
[logical(1)]
Name the result vector? Default is FALSE.

with.nr
[logical(1)]
Should number from 1 to length be appended to name? Only used if use.name and are TRUE. Default is TRUE.

Value

[character].

[Package ParamHelpers version 1.2 Index]