hasFiniteBoxConstraints {ParamHelpers}	R Documentation

Checks if a parameter or each parameter of a parameter set has ONLY finite lower and upper bounds.

Description

Checks if a parameter or each parameter of a parameter set has ONLY finite lower and upper bounds.

Usage

hasFiniteBoxConstraints(par)

Arguments

par
[Param | ParamSet]
Parameter or parameter set.

Value

[logical(1)]

[Package ParamHelpers version 1.2 Index]