isFeasible {ParamHelpers}	R Documentation

Check if parameter value is valid.

Description

Check if a parameter value satisfies the constraints of the parameter description. This includes the requires expressions, if par is a ParamSet. If requires is not satisfied, the parameter value must be set to scalar NA to be still feasible, a single scalar even in a case of a vector parameter.

Usage

  isFeasible(par, x)

Arguments

par
[Param | ParamSet]
Parameter or parameter set.

x
[any]
Single value to check. For a parameter set this must be a list in the correct order.

Value

logical(1)

Examples

p <- makeNumericParam("x", lower=-1, upper=1)
isFeasible(p,0) # True

## [1] TRUE

isFeasible(p,2) # False, out of bounds

## [1] FALSE

isFeasible(p,"a") # False, wrong type

## [1] FALSE

# now for parameter sets
ps <- makeParamSet(
  makeNumericParam("x",lower=-1,upper=1),
  makeDiscreteParam("y", values=c("a","b"))
)
isFeasible(ps, list(0,"a")) # True

## [1] TRUE

isFeasible(ps, list("a",0)) # False, wrong order

## [1] FALSE

[Package ParamHelpers version 1.2 Index]