isForbidden {ParamHelpers}	R Documentation

Check whether parameter setting lies in forbidden region of parameter set.

Description

Parameter sets without a forbidden region always return FALSE.

Usage

isForbidden(par.set, x)

Arguments

par.set
[ParamSet]
Parameter set.

x
[named list]
Parameter setting to check.

Value

[logical(1)].

[Package ParamHelpers version 1.2 Index]