paramValueToString {ParamHelpers}	R Documentation

Convert a value to a string.

Description

Useful helper for logging. For discrete parameter values always the name of the discrete value is used.

Usage

  paramValueToString(par, x, show.missing.values = FALSE)

Arguments

par
[Param | ParamSet]
Parameter or parameter set.

x
[any]
Value for parameter or value for parameter set. In the latter case it must be named list. For discrete parameters their values must be used, not their names.

show.missing.values
[logical(1)]
Display “NA” for parameters, which have no setting, because their requirements are not satified (dependent parameters), instead of displaying nothing? Default is FALSE.

Value

[character(1)].

Examples

p <- makeNumericParam("x")
paramValueToString(p, 1)

## [1] "1.00"

paramValueToString(p, 1.2345)

## [1] "1.23"

p <- makeIntegerVectorParam("x", len=2)
paramValueToString(p, c(1L, 2L))

## [1] "1,2"

p <- makeLogicalParam("x")
paramValueToString(p, TRUE)

## [1] "TRUE"

p <- makeDiscreteParam("x", values=list(a=NULL, b=2))
paramValueToString(p, NULL)

## [1] "a"

ps <- makeParamSet(
  makeNumericVectorParam("x", len=2L),
  makeDiscreteParam("y", values=list(a=NULL, b=2))
)
paramValueToString(ps, list(x=c(1,2), y=NULL))

## [1] "x=1.00,2.00; y=a"

[Package ParamHelpers version 1.2 Index]