Plots attainment functions for data stored in multiple OptPaths.

Can be used to plot OptPaths where information for bi-objective evaluation was logged for repeated runs of different algorithmic runs. Pretty directly calls eaf::eafplot().

plotEAF(opt.paths, xlim = NULL, ylim = NULL, ...)

Arguments

opt.paths
(list)
List of list of OptPath objects. First index is the algorithm / major variation in the experiment, second index is the index of the replicated run.

xlim
(numeric(2))
The x limits (x1, x2) of the plot.

ylim
(numeric(2))
The y limits (y1, y2) of the plot.

...
(any)
Passed on to eaf::eafplot().

Value

data.frame Invisibly returns the data passed to eaf::eafplot().

Note

We changed the defaults of eaf::eafplot() in the following way: The axis are labeled by y.names, colors are set to our favorite grey values and linetypes changed, too. With our colors / linetypes default it is possible to distinguish 6 different algorithms. But this can again be overwritten by the user.