sampleValues {ParamHelpers}	R Documentation

Sample n random values from a parameter or a parameter set uniformly.

Description

Dependent parameters whose requirements are not satisfied are represented by a scalar NA in the output.

Usage

  sampleValues(par, n, discrete.names = FALSE)

Arguments

par
[Param | ParamSet]
Parameter or parameter set.

n
[integer(1)]
Number of values.

discrete.names
[logical(1)]
Should names be sampled for discrete / logical paramaters or values instead? Default is code FALSE.

Value

[list]. For consistency always a list is returned.

Examples

p <- makeIntegerParam("x", lower=-10, upper=10)
sampleValues(p, 4)

## [[1]]
## [1] 9
## 
## [[2]]
## [1] -3
## 
## [[3]]
## [1] 7
## 
## [[4]]
## [1] 9

p <- makeNumericParam("x", lower=-10, upper=10)
sampleValues(p, 4)

## [[1]]
## [1] 5.573
## 
## [[2]]
## [1] -2.392
## 
## [[3]]
## [1] 5.22
## 
## [[4]]
## [1] -7.453

p <- makeLogicalParam("x")
sampleValues(p, 4)

## [[1]]
## [1] TRUE
## 
## [[2]]
## [1] TRUE
## 
## [[3]]
## [1] TRUE
## 
## [[4]]
## [1] FALSE

ps <- makeParamSet(
  makeNumericParam("u", lower=1, upper=10),
  makeIntegerParam("v", lower=1, upper=10),
  makeDiscreteParam("w", values=1:2)
)
sampleValues(ps, 2)

## [[1]]
## [[1]]$u
## [1] 7.827
## 
## [[1]]$v
## [1] 2
## 
## [[1]]$w
## [1] 2
## 
## 
## [[2]]
## [[2]]$u
## [1] 5.644
## 
## [[2]]$v
## [1] 5
## 
## [[2]]$w
## [1] 2

[Package ParamHelpers version 1.2 Index]