setOptPathElEOL {ParamHelpers}	R Documentation

Set the end of life dates of parameter values, in-place.

Description

Set the end of life dates of parameter values, in-place.

Usage

  setOptPathElEOL(op, index, eol)

Arguments

op
[OptPath]
Optimization path.

index
[integer]
Vector of indices of elements.

eol
[integer]
EOL dates, single value or same length of index.

Value

Nothing.

[Package ParamHelpers version 1.2 Index]