generateDesign {ParamHelpers}	R Documentation

Generates a statistical design for a parameter set.

Description

The following types of columns are created:

numeric(vector) 	 numeric

integer(vector) 	 integer

discrete(vector) 	 factor (names of values = levels)

logical(vector) 	 logical

If you want to convert these, look at convertDataFrameCols. Dependent parameters whose constraints are unsatisfied generate NA entries in their respective columns.

Currently only lhs designs are supported.

The algorithm currently iterates the following steps:

We create a space filling design for all parameters, disregarding requires, a trafo or the forbidden region.

Forbidden points are removed.

Parameters are trafoed (maybe); dependent parameters whose constraints are unsatisfied are set to NA entries.

Duplicated design points are removed. Duplicated points are not generated in a reasonable space-filling design, but the way discrete parameters and also parameter dependencies are handled make this possible.

If we removed some points, we now try to augment the design in a space-filling way and iterate.

Note that augmenting currently is somewhat experimental as we simply generate missing points via new calls to randomLHS, but do not add points so they are maximally far away from the already present ones. The reason is that the latter is quite hard to achieve with complicated dependences and forbidden regions, if one wants to ensure that points actually get added... But we are working on it,

Usage

generateDesign(n = 10L, par.set, fun, fun.args = list(), trafo = FALSE,
  augment = 20L)

Arguments

n
[integer(1)]
Number of samples in design. Default is 10.

par.set
[ParamSet]
Parameter set.

fun
[function]
Function from package lhs. Possible are: maximinLHS, randomLHS, geneticLHS, improvedLHS, optAugmentLHS, optimumLHS Default is randomLHS.

fun.args
[list]
List of further arguments passed to fun.

trafo
[logical(1)]
Transform all parameters by using theirs respective transformation functions. Default is FALSE.

augment
[integer(1)]
Duplicated values and forbidden regions in the parameter space can lead to the design becoming smaller than n. With this option it is possible to augment the design again to size n. It is not guaranteed that this always works (to full size) and augment specifies the number of tries to augment. If the the design is of size less than n after all tries, a warning is issued and the smaller design is returned. Default is 20.

Value

The created design is a data.frame. Columns are named by the ids of the parameters. If the par.set argument contains a vector parameter, its corresponding column names in the design are the parameter id concatenated with 1 to dimension of the vector. The result will have an logical(1) attribute “trafo”, which is set to the value of argument trafo.

Examples

ps = makeParamSet(
  makeNumericParam("x1", lower = -2, upper = 1),
  makeIntegerParam("x2", lower = 10, upper = 20)
)
# random latin hypercube design with 5 samples:
generateDesign(5, ps)

## Loading required package: lhs

##        x1 x2
## 1  0.2402 10
## 2 -0.3108 20
## 3  0.6088 12
## 4 -0.8706 15
## 5 -1.5076 17

# with trafo
ps = makeParamSet(
  makeNumericParam("x", lower = -2, upper = 1),
  makeNumericVectorParam("y", len = 2, lower = 0, upper = 1, trafo = function(x) x/sum(x))
)
generateDesign(10, ps, trafo = TRUE)

##           x     y1      y2
## 1  -0.43022 0.5229 0.47711
## 2   0.64060 0.4866 0.51338
## 3  -1.28223 0.1974 0.80262
## 4   0.34936 0.6312 0.36883
## 5  -0.94272 0.4807 0.51934
## 6  -1.54880 0.3263 0.67366
## 7  -0.57284 0.9343 0.06574
## 8  -0.09629 0.1361 0.86387
## 9   0.99170 0.7455 0.25451
## 10 -1.84609 0.2469 0.75306

[Package ParamHelpers version 1.2 Index]