Get lower / upper bounds and allowed discrete values for parameters.

getLower and getUpper return a numerical vector of lower and upper bounds, getValues returns a list of possible value sets for discrete parameters.

Parameters for which such bound make no sense - due to their type - are not present in the result.

getLower(obj, with.nr = FALSE, dict = NULL)

getUpper(obj, with.nr = FALSE, dict = NULL)

getValues(obj, dict = NULL)

Arguments

obj
(Param() | ParamSet() | list)
Parameter, parameter set or list of parameters, whose boundaries and/or values should be extracted. In case the boundaries or values contain expressions, they will be evaluated using the provided dictionary dict.

with.nr
(logical(1))
Should number from 1 to length be appended to names of vector params? Default is FALSE.

dict
(environment | list | NULL)
Environment or list which will be used for evaluating the variables of expressions within a parameter, parameter set or list of parameters. The default is NULL.

Value

vector | list. Named by parameter ids.

Examples

ps = makeParamSet(
  makeNumericParam("u"),
  makeDiscreteParam("v", values = c("a", "b")),
  makeIntegerParam("w", lower = expression(ceiling(p / 3)), upper = 2),
  makeDiscreteParam("x", values = 1:2),
  makeNumericVectorParam("y", len = 2, lower = c(0, 10), upper = c(1, 11)),
  keys = "p"
)
getLower(ps, dict = list(p = 7))

#>    u    w    y    y 
#> -Inf    3    0   10 

getUpper(ps)

#>   u   w   y   y 
#> Inf   2   1  11 

ps = makeParamSet(
  makeNumericParam("u"),
  makeDiscreteParam("w", values = list(a = list(), b = NULL))
)
getValues(ps)

#> $w
#> $w$a
#> list()
#> 
#> $w$b
#> NULL
#> 
#> 

par.vals = list(
  u = makeNumericParam("u"),
  v = makeIntegerParam("v", lower = 1, upper = 2),
  w = makeDiscreteParam("w", values = 1:2),
  x = makeNumericVectorParam("x", len = 2, lower = c(3, 1), upper = expression(n))
)
getLower(par.vals)

#>    u    v    x    x 
#> -Inf    1    3    1 

getUpper(par.vals, dict = list(n = 12))

#>   u   v   x   x 
#> Inf   2  12  12