makeParamSet {ParamHelpers}	R Documentation

Construct a parameter set.

Description

makeParamSet: Contruct from a bunch of parameters.

Multiple sets can be concatenated with c.

The constructed S3 class is simply a list that contains the element pars. pars is a list of the passed parameters, named by their ids.

makeNumericParamSet: Convenience function for numerics.

Usage

makeParamSet(..., params, forbidden = NULL)

makeNumericParamSet(id = "x", len, lower = -Inf, upper = Inf,
  vector = TRUE)

Arguments

...
[Param]
Parameters.

params
[list of Param]
List of parameters, alternative way instead of using ....

forbidden
[NULL | R expression]
States forbidden region of parameter set via an expression. Every setting which satisfies this expression is considered to be infeasible. This makes it possible to exclude more complex region of the parameter space than through simple constraints or requires-conditions (although these should be always used when possible). If parameters have associated trafos, the forbidden region must always be specified on the original scale and not the transformed one. Default is NULL which means no forbidden region.

id
[character(1)] Name of parameter.

len
[integer(1)]
Length of vector.

lower
[numeric]
Lower bound. Default is -Inf.

upper
[numeric]
Upper bound. Default is Inf.

vector
[logical(1)]
Should a NumericVectorParam be used instead of n NumericParam objects? Default is TRUE.

Value

[ParamSet].

Examples

makeParamSet(
  makeNumericParam("u", lower=1),
  makeIntegerParam("v", lower=1, upper=2),
  makeDiscreteParam("w", values=1:2),
  makeLogicalParam("x"),
  makeDiscreteVectorParam("y", len=2, values=c("a", "b"))
)

## Warning: invalid factor level, NA generated

##             Type  len Def   Constr Req Trafo
## u        numeric    -   - 1 to Inf   -     -
## v        integer    -   -   1 to 2   -     -
## w       discrete    -   -      1,2   -     -
## x        logical    -   -        -   -     -
## y discretevector <NA>   -      a,b   -     -

[Package ParamHelpers version 1.2 Index]