sampleValues {ParamHelpers}	R Documentation

Dependent parameters whose requirements are not satisfied are represented by a scalar NA in the output.

Description

Dependent parameters whose requirements are not satisfied are represented by a scalar NA in the output.

Usage

sampleValues(par, n, discrete.names = FALSE, trafo = FALSE)

Arguments

par
[Param | ParamSet]
Parameter or parameter set.

n
[integer(1)]
Number of values.

discrete.names
[logical(1)]
Should names be sampled for discrete / logical paramaters or values instead? Default is code FALSE.

trafo
[logical(1)]
Transform all parameters by using theirs respective transformation functions. Default is FALSE.

Value

[list]. For consistency always a list is returned.

Examples

p = makeIntegerParam("x", lower = -10, upper = 10)
sampleValues(p, 4)

## [[1]]
## [1] 8
## 
## [[2]]
## [1] -5
## 
## [[3]]
## [1] -7
## 
## [[4]]
## [1] 9

p = makeNumericParam("x", lower = -10, upper = 10)
sampleValues(p, 4)

## [[1]]
## [1] -4.652
## 
## [[2]]
## [1] -3.236
## 
## [[3]]
## [1] 0.8555
## 
## [[4]]
## [1] -6.149

p = makeLogicalParam("x")
sampleValues(p, 4)

## [[1]]
## [1] FALSE
## 
## [[2]]
## [1] FALSE
## 
## [[3]]
## [1] TRUE
## 
## [[4]]
## [1] FALSE

ps = makeParamSet(
  makeNumericParam("u", lower = 1, upper = 10),
  makeIntegerParam("v", lower = 1, upper = 10),
  makeDiscreteParam("w", values = 1:2)
)
sampleValues(ps, 2)

## [[1]]
## [[1]]$u
## [1] 9.538
## 
## [[1]]$v
## [1] 4
## 
## [[1]]$w
## [1] 1
## 
## 
## [[2]]
## [[2]]$u
## [1] 6.073
## 
## [[2]]$v
## [1] 3
## 
## [[2]]$w
## [1] 1

[Package ParamHelpers version 1.2 Index]