setOptPathElDOB {ParamHelpers}	R Documentation

Set the dates of birth of parameter values, in-place.

Description

Set the dates of birth of parameter values, in-place.

Usage

  setOptPathElDOB(op, index, dob)

Arguments

op
[OptPath]
Optimization path.

index
[integer]
Vector of indices of elements.

dob
[integer]
Dates of birth, single value or same length of index.

Value

Nothing.

[Package ParamHelpers version 1.2 Index]