Transform optimization path.

Transform optimization path with associated transformation functions of parameters. Can only be done when x values where added “untransformed”.

trafoOptPath(opt.path)

Arguments

opt.path

[OptPath()]
Optimization path.

Value

[OptPath()].

Examples

ps = makeParamSet(
  makeIntegerParam("u", trafo = function(x) 2 * x),
  makeNumericVectorParam("v", len = 2, trafo = function(x) x / sum(x)),
  makeDiscreteParam("w", values = c("a", "b"))
)
op = makeOptPathDF(ps, y.names = "y", minimize = TRUE)
addOptPathEl(op, x = list(3, c(2, 4), "a"), y = 0, dob = 1, eol = 1)
addOptPathEl(op, x = list(4, c(5, 3), "b"), y = 2, dob = 5, eol = 7)

as.data.frame(op)
#>   u v1 v2 w y dob eol
#> 1 3  2  4 a 0   1   1
#> 2 4  5  3 b 2   5   7
op = trafoOptPath(op)
as.data.frame(op)
#>   u        v1        v2 w y dob eol
#> 1 6 0.3333333 0.6666667 a 0   1   1
#> 2 8 0.6250000 0.3750000 b 2   5   7